v1.0.0

+++ model/Cards/Pineco.java
@@ -6,10 +6,10 @@ import enums.CardSellPrices;
 import model.Game;
 import model.Player;
 
-public class Pineco extends Pokemon{
+public class Pineco extends Pokemon {
     public Pineco() {
         super(CardBuyPrices.Pineco.price, CardSellPrices.Pineco.price, CardNames.Pineco.name, "plant",
-                110,25,0.9,1,1,1,15);
+                110, 25, 0.9, 1, 1, 1, 15);
     }
 
     public void doPassive(Game game, Player player, Player enemy) {
@@ -18,7 +18,7 @@ public class Pineco extends Pokemon{
 
     public void doAction(Game game, Player player, Player enemy) {
         Pokemon playerPokemon = player.getActiveCard();
-        int heal = (int) playerPokemon.getDamageToTarget(playerPokemon);
-        playerPokemon.setHitpoint(playerPokemon.getHitpoint()+heal);
+        double heal = playerPokemon.getDamageToTarget(playerPokemon);
+        playerPokemon.setHitpoint(playerPokemon.getHitpoint() + heal);
     }
 }
